requirements.txt
black==22.3.0
fastapi==0.110.2
pylint==2.13.5
python-dotenv==1.0.1
pydantic==2.7.1
pydantic_settings
PyYAML==5.3.1
pytest==7.1.1
makefun==1.13.1
mypy==0.942
uvicorn==0.29.0
